Use listener on pages/cleanup
