+++ ui/page/staking/list_page.go
@@ -9,12 +9,12 @@ import (
 	"gioui.org/widget"
 
 	"github.com/planetdecred/dcrlibwallet"
+	"github.com/planetdecred/godcr/listeners"
 	"github.com/planetdecred/godcr/ui/decredmaterial"
 	"github.com/planetdecred/godcr/ui/load"
 	"github.com/planetdecred/godcr/ui/page/components"
 	tpage "github.com/planetdecred/godcr/ui/page/transaction"
 	"github.com/planetdecred/godcr/ui/values"
-	"github.com/planetdecred/godcr/wallet"
 )
 
 const listPageID = "StakingList"
@@ -33,7 +33,7 @@ const (
 
 type ListPage struct {
 	*load.Load
-
+	*listeners.TxAndBlockNotification
 	ctx       context.Context // page context
 	ctxCancel context.CancelFunc
 
@@ -87,6 +87,13 @@ func (pg *ListPage) ID() string {
 // the page is displayed.
 // Part of the load.Page interface.
 func (pg *ListPage) OnNavigatedTo() {
+	if pg.TxAndBlockNotification == nil {
+		pg.TxAndBlockNotification = listeners.NewTxAndBlockNotification(make(chan listeners.TxNotification, 4))
+	} else {
+		pg.TxAndBlockNotifCh = make(chan listeners.TxNotification, 4)
+	}
+	pg.WL.MultiWallet.AddTxAndBlockNotificationListener(pg.TxAndBlockNotification, true, listPageID)
+
 	pg.ctx, pg.ctxCancel = context.WithCancel(context.TODO())
 	pg.listenForTxNotifications()
 	pg.fetchTickets()
@@ -95,27 +102,25 @@ func (pg *ListPage) OnNavigatedTo() {
 func (pg *ListPage) listenForTxNotifications() {
 	go func() {
 		for {
-			var notification interface{}
 
 			select {
-			case notification = <-pg.Receiver.NotificationsUpdate:
+			case n := <-pg.TxAndBlockNotifCh:
+				if n.NotificationType == listeners.BlkAttached {
+					selectedWallet := pg.wallets[pg.walletDropDown.SelectedIndex()]
+					if selectedWallet.ID == n.WalletID {
+						pg.fetchTickets()
+						pg.RefreshWindow()
+					}
+				} else if n.NotificationType == listeners.NewTx {
+					pg.UpdateBalance()
+					selectedWallet := pg.wallets[pg.walletDropDown.SelectedIndex()]
+					if selectedWallet.ID == n.Transaction.WalletID {
+						pg.fetchTickets()
+						pg.RefreshWindow()
+					}
+				}
 			case <-pg.ctx.Done():
 				return
-			}
-
-			switch n := notification.(type) {
-			case wallet.NewBlock:
-				selectedWallet := pg.wallets[pg.walletDropDown.SelectedIndex()]
-				if selectedWallet.ID == n.WalletID {
-					pg.fetchTickets()
-					pg.RefreshWindow()
-				}
-			case wallet.NewTransaction:
-				selectedWallet := pg.wallets[pg.walletDropDown.SelectedIndex()]
-				if selectedWallet.ID == n.Transaction.WalletID {
-					pg.fetchTickets()
-					pg.RefreshWindow()
-				}
 			}
 		}
 	}()
@@ -289,4 +294,8 @@ func (pg *ListPage) HandleUserInteractions() {
 // Part of the load.Page interface.
 func (pg *ListPage) OnNavigatedFrom() {
 	pg.ctxCancel()
+	if pg.TxAndBlockNotifCh != nil {
+		close(pg.TxAndBlockNotifCh)
+	}
+	pg.WL.MultiWallet.RemoveTxAndBlockNotificationListener(listPageID)
 }
